LG Encryption working with some barbaric edits
1. We should not symlink any files to /vendor. We need those files.
2. Since we are using own own services we should not set any properties.
3. Force use our own vdc. vdc_pie doesnt have the command we need.
4. Trigger normal default decryption used while booting Android
Unfortunately if we run "checkpw" on LG's vold it will return that we have the correct password however it will not set ro.crypto.fs_crypto_blkdev at all (which we need to mount data). Instead we can run "mountdefaultencrypted" to start normal decryption process. The only caviats is that some vold props must be set manually at some point and /data will be mounted by vold instead of TWRP.

Change-Id: Id421ee0d75e74314fda4a29a2716803ded0cbba8